Fix caching
PWA was failing to work in offline mode due to not properly caching its
resources. Turns out, it was `vite-plugin-pwa` passing in objects and
not URLs for the precache resources. It might be workbox doing it but
who knows, as the plugin authors just shell out to workbox to do mostly
anything.