Fix off-by-one error
Temporal has 1-based week-day numbers while normal arrays use 0-based.

Also, apparently Monday is the first day of the week and not Sunday. I'm
accustomed to Sunday being the first but the ISO standard uses Monday. I
could rewrite the logic to make Sunday the first but can't be bothered
at the moment.